pyproject.toml
[project]
name = "nic"
version = "0.1.0"
description = "Hierarchical image clustering API for cake images"
readme = "README.md"
authors = [
    { name = "masa-57", email = "m.ahanchian@outlook.com" }
]
requires-python = ">=3.12,<3.13"
dependencies = [
    # API
    "fastapi[standard]>=0.115",
    "python-multipart>=0.0.9",
    # Hashing / image utilities (lightweight, used by API search + worker)
    "imagehash>=4.3",
    "Pillow>=10.0",
    "numpy>=1.26",
    # Vector Store (pgvector)
    "pgvector>=0.3",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    # AWS
    "boto3>=1.35",
    # Config
    "pydantic-settings>=2.0",
    "modal>=1.3.2",
    # Observability
    "prometheus-fastapi-instrumentator>=7.0",
    "sentry-sdk[fastapi,sqlalchemy]>=2.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # Retry logic
    "tenacity>=8.0",
    # Google Drive Sync
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
]

[project.optional-dependencies]
ml = [
    # ML / Embeddings (worker-only, ~2GB)
    "torch>=2.2",
    "transformers>=4.40",
    # Clustering
    "scikit-learn>=1.4",
    "umap-learn>=0.5.7",
    "numba>=0.59",
    "scipy>=1.12",
]

[project.scripts]
nic = "nic.main:app"

[build-system]
requires = ["uv_build>=0.9.22,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.5",
    "mypy>=1.10",
    "pytest-cov>=5.0",
    "coverage>=7.0",
    "testcontainers[postgres]>=4.0",
    "moto[s3]>=5.0",
    "psycopg2-binary>=2.9",
    "pip-audit>=2.7",
]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src"]
extend-exclude = ["src/nic/migrations"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH", "C901", "S", "TRY", "PT", "RET", "PIE"]
ignore = [
    "B008",    # FastAPI Depends/Query in defaults
    "UP047",   # TypeVar clearer than PEP 695 for generic helpers
    "TRY003",  # Long exception messages OK in validators/config
]

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]           # assert in tests is fine
"scripts/**/*.py" = ["S310", "S603"] # scripts use urllib + subprocess legitimately

[tool.ruff.lint.isort]
known-first-party = ["nic"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "unit: fast unit tests (no external deps)",
    "integration: requires PostgreSQL (testcontainers)",
    "e2e: full pipeline tests (slow, requires model)",
]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/nic"]

[tool.coverage.report]
fail_under = 70

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
